DBC22-3525: markAdvisoriesAsRead now a shared function
